Use uint16 instead of uint32 for discv5 log distance

+++ eth/p2p/discoveryv5/messages.nim
@@ -47,7 +47,7 @@ type
     port*: uint16
 
   FindNodeMessage* = object
-    distances*: seq[uint32]
+    distances*: seq[uint16]
 
   NodesMessage* = object
     total*: uint32
